% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{cb}
\alias{cb}
\title{Cudeck & Browne (1992) model error method}
\usage{
cb(target_rmsea, mod)
}
\arguments{
\item{target_rmsea}{(scalar) Target RMSEA value.}

\item{mod}{A `fungible::simFA()` model object.}
}
\value{

}
\description{
Generate a population correlation matrix using the model described in Cudeck
and Browne (1992).
}
\examples{
set.seed(42)
mod <- fungible::simFA()
Sigma <- cb(
  target_rmsea = 0.05,
  mod = mod
)
# Verify the result
rmsea(Sigma, mod$Rpop, k = ncol(mod$loadings))
}
\references{
Cudeck, R., & Browne, M. W. (1992). Constructing a covariance matrix that yields a specified minimizer and a specified minimum discrepancy function value. \emph{Psychometrika}, \emph{57}(3), 357–369. \url{https://doi.org/10/cq6ckd}
}
