% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi.R
\name{cfi}
\alias{cfi}
\title{Calculate CFI for two correlation matrices}
\usage{
cfi(Sigma, Omega)
}
\arguments{
\item{Sigma}{(matrix) Population correlation or covariance matrix (with model
error).}

\item{Omega}{(matrix) Model-implied population correlation or covariance
matrix.}
}
\value{

}
\description{
Given two correlation matrices of the same dimension, calculate the CFI value
value using the independence model as the null model.
}
\examples{
set.seed(42)
mod <- fungible::simFA(Model = list(NFac = 3))
Omega <- mod$Rpop
Sigma <- cb(
  mod = mod,
  target_rmsea = 0.05
)
cfi(Sigma, Omega)
}
