% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wb_coef.R
\name{find_wb_coef}
\alias{find_wb_coef}
\title{Find a cofficient to use with the Wu & Browne (2015) model error method}
\usage{
find_wb_coef(mod, n = 250, values = 2, lower = 0.01, upper = 0.1)
}
\arguments{
\item{mod}{A \code{fungible::simFA()} model object.}

\item{n}{The number of times to evaluate \code{wb()} at each point.}

\item{values}{The number of target RMSEA values to evaluate between 0.02 and
0.1.}

\item{lower}{(scalar) The smallest target RMSEA value to use.}

\item{upper}{(scalar) The largest target RMSEA value to use.}
}
\value{
(scalar) A coefficient to use with the \code{\link{wb}} function to
obtain population correlation matrices with model error that have RMSEA
values closer to the target RMSEA values.
}
\description{
The Wu & Browne (2015) model error method takes advantage of the relationship
between v and RMSEA:
}
\details{
\deqn{v = RMSEA^2 + o(RMSEA^2).}

As RMSEA increases, the approximation \eqn{v ~= RMSEA^2} becomes worse. This
function generates population correlation matrices with model error for
multiple target RMSEA values and then regresses the median (observed) RMSEA
values on the target RMSEA values. The reciprocal of the regression
coefficient is returned and can be used to adjust the value of \eqn{v} in
calls to \code{wb()} to ensure that generated matrices have RMSEA values that are
close to the target value.
}
\examples{
set.seed(42)
mod <- fungible::simFA(Seed = 42)
wb_coef <- find_wb_coef(mod, n = 100, values = 5,
                        lower = .045, upper = .055)
wb(mod$Rpop, target_rmsea = 0.05, wb_coef = wb_coef)
}
