% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsea.R
\name{rmsea}
\alias{rmsea}
\title{Calculate RMSEA between two correlation/covariance matrices}
\usage{
rmsea(Sigma, Omega, k)
}
\arguments{
\item{Sigma}{(matrix) Population correlation or covariance matrix (with model error).}

\item{Omega}{(matrix) Model-implied population correlation or covariance matrix.}

\item{k}{(scalar) Number of major common factors.}

\item{df}{(scalar) Model degrees of freedom.}
}
\value{

}
\description{
Calculate RMSEA between two correlation/covariance matrices
}
\examples{
set.seed(42)
mod <- fungible::simFA()
Omega <- mod$Rpop
Sigma <- wu_browne(target_rmsea = 0.05,
                   Omega = Omega)
rmsea(Sigma, Omega)
}
