% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkl.R
\name{tkl}
\alias{tkl}
\title{Optimize TKL parameters to find a solution with target RMSEA and CFI values}
\usage{
tkl(
  L,
  Phi,
  target_rmsea,
  target_cfi,
  weights = c(1, 1),
  v_start = 0.1,
  eps_start = 0.02,
  NMinorFac = 50,
  ModelErrorType = "U",
  WmaxLoading = NULL,
  NWmaxLoading = 2,
  debug = FALSE,
  penalty = 500,
  optim_type = "ga",
  seed = NULL,
  ncores = FALSE
)
}
\arguments{
\item{L}{(matrix) Matrix of major common factor loadings.}

\item{Phi}{(matrix) Major common factor correlation matrix.}

\item{target_rmsea}{(scalar) Target RMSEA value.}

\item{target_cfi}{(scalar) Target CFI value.}

\item{weights}{(vector) Vector of length two indicating how much weight to
give RMSEA and CFI, e.g., `c(1,1)` (default) gives equal weight to both
indices; `c(1,0)` ignores the CFI value.}

\item{v_start}{(scalar) Starting value to use for \eqn{\upsilon}, the
proportion of uniqueness variance reallocated to the minor common factors.}

\item{eps_start}{(scalar) Starting value to use for \eqn{\epsilon}, which
controls how common variance is distributed among the minor common factors.}

\item{ModelErrorType}{(character) "U" or "V", as defined in `simFA()`.}

\item{WmaxLoading}{(scalar) Threshold value for `NWmaxLoading`.}

\item{NWmaxLoading}{(scalar) Maximum number of absolute loadings \eqn{\ge}
`WmaxLoading` in any column of \eqn{W}.}

\item{penalty}{(scalar) Penalty applied to objective function if the
NmaxLoading condition isn't satisfied.}

\item{optim_type}{(character)  Which optimization function to use, `optim` or
`ga`? `optim()` is faster, but might not converge in some cases.}

\item{seed}{(scalar) seed for reproducibility.}

\item{ncores}{(boolean/scalar) Controls whether `ga()` optimization is done in
parallel. If `TRUE`, uses the maximum available number of processor cores. If
`FALSE`, does not use parallel processing. If an integer is provided, that's
how many processor cores will be used (if available).}

\item{W}{(matrix) Matrix of provisional minor common factor loadings with
unit column variances.}

\item{NminorFac}{(scalar) Number of minor common factors.}
}
\description{
Find the optimal W matrix such that the RMSEA and CFI values are as close as
possible to the user-specified target values.
}
\examples{
mod <- fungible::simFA()
tkl_opt(
  L = mod$loadings,
  Phi = mod$Phi,
  target_rmsea = 0.05,
  target_cfi = 0.95,
  WmaxLoading = .35,
  NWmaxLoading = 3,
  optim_type = "optim",
  seed = 123
)
}
