% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\name{wb}
\alias{wb}
\title{Wu & Browne model error method}
\usage{
wb(target_rmsea, Omega)
}
\arguments{
\item{target_rmsea}{(scalar) Target RMSEA value.}

\item{Omega}{(matrix) Model-implied population correlation matrix.}
}
\value{

}
\description{
Generate a population correlation matrix using the model described in Wu and
Browne (2015).
}
\examples{
set.seed(42)
R <- fungible::simFA()$Rpop
wb(
  target_rmsea = 0.05,
  Omega = R
)
}
\references{
Wu, H., & Browne, M. W. (2015). Quantifying adventitious error in a covariance structure as a random effect. \emph{Psychometrika}, \emph{80}(3), 571–600. \url{https://doi.org/10/gjrkc4}
}
\author{
Justin Kracht <krach018@umn.edu>
}
